% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_density.R
\name{color_by_density}
\alias{color_by_density}
\title{Color Images by density}
\usage{
color_by_density(new_name, image_path, palette, mode = "grey")
}
\arguments{
\item{new_name}{the new name for the file}

\item{image_path}{the path to the source image}

\item{palette}{the new palette to apply}

\item{mode}{A character value specifying target mode for conversion using EBIImage See details}
}
\value{
nothing
}
\description{
I think this code is adapted from a blog post by
Will R Chase. This takes an image and maps new colors by density
then saves the image with the new name.
This uses EBImage from Bioconductor.
}
\details{
Conversion Modes:
\itemize{
\item \code{"rgb"}: Converts a Grayscale image or an array into a Color image, replicating RGB channels.
\item \verb{"gray" or "grey"} Converts a Color image into a Grayscale image, using uniform 1/3 RGB weights.
\item \verb{red, green, blue}: Extracts the red, green or blue channel from a Color image. Returns a Grayscale image.
\item \verb{asred, asgreen, asblue}: Converts a Grayscale image or an array into a Color image of the specified hue.
}
}
