% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_paint.R
\name{get_random_pal}
\alias{get_random_pal}
\title{Select a random palette from a list}
\usage{
get_random_pal(pal_list = delabjGenArt::paint_palettes, seed = NULL)
}
\arguments{
\item{pal_list}{a named list of palettes}

\item{seed}{optional, select a seed for consistent results.}
}
\value{
a character vector with hex codes.
}
\description{
A helper function to select random palettes for systems
}
