% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_shapes.R
\name{make_line}
\alias{make_line}
\title{Interpolate points between two nodes}
\usage{
make_line(n_points, x_start, y_start, x_end, y_end, ...)
}
\arguments{
\item{n_points}{how many points are drawn between the nodes?}

\item{x_start}{starting x}

\item{y_start}{starting y}

\item{x_end}{ending x}

\item{y_end}{ending y}

\item{...}{discarded}
}
\description{
Interpolate points between two nodes
}
