% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_paint.R
\docType{data}
\name{paint_palettes}
\alias{paint_palettes}
\title{Generative Palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
paint_palettes
}
\description{
A collection of palettes to be used in generative work
}
\keyword{datasets}
