% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_points.R
\name{rotate_points}
\alias{rotate_points}
\title{Rotate points about}
\usage{
rotate_points(data, x, y, deg = 0, cx = 0, cy = 0)
}
\arguments{
\item{data}{a data.frame}

\item{x}{column with x coordinates}

\item{y}{column with y coordinates}

\item{deg}{the degree to rotate the points}

\item{cx}{the x coordinate to rotate around}

\item{cy}{the y coordinate to rotate around}
}
\value{
a data.frame
}
\description{
take a set of and rotate it in degrees
}
