% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_canvas}
\alias{theme_canvas}
\title{A No Frills theme for generative art}
\usage{
theme_canvas(background_fill = "white")
}
\arguments{
\item{background_fill}{color of the plot background

# return A list with elements to modify a ggplot object}
}
\description{
A No Frills theme for generative art
}
